+++ src/components/life/solo-life-list.tsx
@@ -1,6 +1,6 @@
-import { SoloLifeCardType } from '@/types/solo-life';
 import SoloLifeCard from '@/components/life/solo-life-card';
 import { useLifePostsByMonth } from '@/lib/mutations/useLifePostsByMonth';
+import { SoloLifeCardType } from '@/types/life-post';
 
 const SoloLifeList = ({ selectedDate }: { selectedDate: string }) => {
   const selectedMonth = selectedDate.slice(0, 7); // '2025-04'
